requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
click==8.1.8
easyocr==1.7.2
fastapi==0.115.8
filelock==3.17.0
fsspec==2025.2.0
h11==0.14.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.5
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.3
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
python-bidi==0.6.6
python-multipart==0.0.20
PyYAML==6.0.2
scikit-image==0.25.2
scipy==1.15.2
setuptools==75.8.0
shapely==2.0.7
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
tifffile==2025.2.18
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.12.2
uvicorn==0.34.0
